+++ hltas-cpp-bridge/src/x86_64-unknown-linux-gnu.rs
@@ -113,6 +113,8 @@ pub enum StrafeDir {
     YAW = 3,
     POINT = 4,
     LINE = 5,
+    LEFT_RIGHT = 6,
+    RIGHT_LEFT = 7,
 }
 #[repr(u8)]
 #[derive(Debug, Copy, Clone, Hash, PartialEq, Eq)]
@@ -598,6 +600,7 @@ pub struct hltas_frame {
     pub Yaw: f64,
     pub X: f64,
     pub Y: f64,
+    pub Count: ::std::os::raw::c_uint,
     pub Pitch: f64,
     pub Repeats: u32,
     pub Commands: *const ::std::os::raw::c_char,
@@ -626,7 +629,7 @@ pub struct hltas_frame {
 fn bindgen_test_layout_hltas_frame() {
     assert_eq!(
         ::std::mem::size_of::<hltas_frame>(),
-        248usize,
+        256usize,
         concat!("Size of: ", stringify!(hltas_frame))
     );
     assert_eq!(
@@ -1015,8 +1018,18 @@ fn bindgen_test_layout_hltas_frame() {
         )
     );
     assert_eq!(
-        unsafe { &(*(::std::ptr::null::<hltas_frame>())).Pitch as *const _ as usize },
+        unsafe { &(*(::std::ptr::null::<hltas_frame>())).Count as *const _ as usize },
         104usize,
+        concat!(
+            "Offset of field: ",
+            stringify!(hltas_frame),
+            "::",
+            stringify!(Count)
+        )
+    );
+    assert_eq!(
+        unsafe { &(*(::std::ptr::null::<hltas_frame>())).Pitch as *const _ as usize },
+        112usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1026,7 +1039,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).Repeats as *const _ as usize },
-        112usize,
+        120usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1036,7 +1049,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).Commands as *const _ as usize },
-        120usize,
+        128usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1046,7 +1059,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).Comments as *const _ as usize },
-        128usize,
+        136usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1056,7 +1069,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).SaveName as *const _ as usize },
-        136usize,
+        144usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1066,7 +1079,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).SeedPresent as *const _ as usize },
-        144usize,
+        152usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1076,7 +1089,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).Seed as *const _ as usize },
-        148usize,
+        156usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1086,7 +1099,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).BtnState as *const _ as usize },
-        152usize,
+        160usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1096,7 +1109,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).Buttons as *const _ as usize },
-        153usize,
+        161usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1108,7 +1121,7 @@ fn bindgen_test_layout_hltas_frame() {
         unsafe {
             &(*(::std::ptr::null::<hltas_frame>())).LgagstMinSpeedPresent as *const _ as usize
         },
-        157usize,
+        165usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1118,7 +1131,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).LgagstMinSpeed as *const _ as usize },
-        160usize,
+        168usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1128,7 +1141,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).ResetFrame as *const _ as usize },
-        164usize,
+        172usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1140,7 +1153,7 @@ fn bindgen_test_layout_hltas_frame() {
         unsafe {
             &(*(::std::ptr::null::<hltas_frame>())).ResetNonSharedRNGSeed as *const _ as usize
         },
-        168usize,
+        176usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1152,7 +1165,7 @@ fn bindgen_test_layout_hltas_frame() {
         unsafe {
             &(*(::std::ptr::null::<hltas_frame>())).StrafingAlgorithmPresent as *const _ as usize
         },
-        176usize,
+        184usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1162,7 +1175,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).Algorithm as *const _ as usize },
-        180usize,
+        188usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1174,7 +1187,7 @@ fn bindgen_test_layout_hltas_frame() {
         unsafe {
             &(*(::std::ptr::null::<hltas_frame>())).AlgorithmParametersPresent as *const _ as usize
         },
-        184usize,
+        192usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1184,7 +1197,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).Parameters as *const _ as usize },
-        192usize,
+        200usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1194,7 +1207,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).ChangePresent as *const _ as usize },
-        216usize,
+        224usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1204,7 +1217,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).Target as *const _ as usize },
-        217usize,
+        225usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1214,7 +1227,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).ChangeFinalValue as *const _ as usize },
-        220usize,
+        228usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1224,7 +1237,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).ChangeOver as *const _ as usize },
-        224usize,
+        232usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1234,7 +1247,7 @@ fn bindgen_test_layout_hltas_frame() {
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<hltas_frame>())).TargetYawOverride as *const _ as usize },
-        232usize,
+        240usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),
@@ -1246,7 +1259,7 @@ fn bindgen_test_layout_hltas_frame() {
         unsafe {
             &(*(::std::ptr::null::<hltas_frame>())).TargetYawOverrideCount as *const _ as usize
         },
-        240usize,
+        248usize,
         concat!(
             "Offset of field: ",
             stringify!(hltas_frame),