+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Added
+- `StrafeDir::LeftRight`, `StrafeDir::RightLeft`.
 - cpp: `Frame::HasYaw()`, `Frame::HasXY()`.
 
 ## [0.5.0] - 31 Dec 2021